Move entity types to types/entity
