+++ src/lib/components/Header.svelte
@@ -24,7 +24,8 @@ The header (navigation bar) of the application.
 	import { page } from "$app/state";
 	import { OIDC_STATE_KEY } from "$lib";
 	import { Title } from "$lib/components";
-	import type { AuthInfo, User } from "$lib/types";
+	import type { AuthInfo } from "$lib/types";
+	import type { User } from "$lib/types/entity";
 	import { innerWidth } from "svelte/reactivity/window";
 
 	type Props = {
