+++ src/routes/admin/users/[userId]/+layout.server.ts
@@ -1,6 +1,6 @@
 import { error, isHttpError } from "@sveltejs/kit";
 import type { LayoutServerLoad } from "./$types";
-import type { UserWithRoles } from "$lib/types";
+import type { UserWithRoles } from "$lib/types/entity";
 
 export const load = (async (event) => {
 	try {
